Added files via upload
Please not that If you exit the application the alarms will continue to execute. When an alarm is ringing a speaker is displayed in the notification bar. You can click and drag it down to view it. If you click on it, it will take you to the alarm activity to shut the alarm off. 
Thank you.